+++ src/Eloquent/Schemer/Loader/Exception/InvalidUriTypeException.php
@@ -13,23 +13,28 @@ namespace Eloquent\Schemer\Loader\Exception;
 
 use Eloquent\Schemer\Uri\UriInterface;
 use Exception;
-use RuntimeException;
+use LogicException;
 
-final class LoadException extends RuntimeException implements
-    LoadExceptionInterface
+final class InvalidUriTypeException extends LogicException
 {
     /**
      * @param UriInterface   $uri
+     * @param string         $expectedClass
      * @param Exception|null $previous
      */
-    public function __construct(UriInterface $uri, Exception $previous = null)
-    {
+    public function __construct(
+        UriInterface $uri,
+        $expectedClass,
+        Exception $previous = null
+    ) {
         $this->uri = $uri;
+        $this->expectedClass = $expectedClass;
 
         parent::__construct(
             sprintf(
-                'Unable to read from %s.',
-                var_export($uri->toString(), true)
+                'Invalid URI of type %s passed. Instance of %s expected',
+                var_export(get_class($uri), true)
+                var_export($expectedClass, true)
             ),
             0,
             $previous
@@ -44,5 +49,14 @@ final class LoadException extends RuntimeException implements
         return $this->uri;
     }
 
+    /**
+     * @return string
+     */
+    public function expectedClass()
+    {
+        return $this->expectedClass;
+    }
+
     private $uri;
+    private $expectedClass;
 }