+++ src/Eloquent/Schemer/Reference/SwitchingResolutionScopeMapFactory.php
@@ -19,8 +19,8 @@ use Eloquent\Schemer\Uri\Resolver\UriResolver;
 use Eloquent\Schemer\Uri\Resolver\UriResolverInterface;
 use Eloquent\Schemer\Uri\UriFactory;
 use Eloquent\Schemer\Uri\UriFactoryInterface;
+use Eloquent\Schemer\Uri\UriInterface;
 use Eloquent\Schemer\Value;
-use Zend\Uri\UriInterface;
 
 class SwitchingResolutionScopeMapFactory extends Value\Visitor\AbstractValueVisitor implements
     ResolutionScopeMapFactoryInterface